Additions to SConscript files to build and run analysis locally on OS X.
git-svn-id: https://repo.eecs.berkeley.edu/svn/users/ksen/projects/corvette-instrumentor@6766 1c3fd406-9068-4342-8353-511fcdf43622

+++ src/InterpreterObserver.cpp
@@ -1563,7 +1563,7 @@ void InterpreterObserver::getelementptr(IID iid UNUSED, bool inbound UNUSED, KVA
   index = index + basePtrLocation->getIndex();
 
   safe_assert(index < (int) basePtrLocation->getLength());
-  if (index < 0 || (int) array[index].getFirstByte() != index*size/8 || KIND_GetSize(array[index].getType()) != (int) size/8) {
+  if (index < 0 || (int) array[index].getFirstByte() != index*(int)size/8 || KIND_GetSize(array[index].getType()) != (int) size/8) {
     index = findIndex((IValue*) basePtrLocation->getIPtrValue(), newOffset,
         basePtrLocation->getLength()); 
   }
